refactor(cli): Extract shared open_keyring function
The keyring-opening logic with backend selection and auto-detect
was duplicated across 4 files with inconsistent path resolution.
Extract a single open_keyring() into commands/mod.rs and call it
from all sites.
