+++ crates/cli/src/commands/start/p2p.rs
@@ -13,72 +13,9 @@ impl Node {
     /// If a peer key exists in the keyring, it is loaded and converted to a libp2p Keypair.
     /// If no peer key exists, a new Ed25519 key is generated and stored in the keyring.
     pub(super) fn init_peer_key(config: &Config) -> Result<(p2p::Keypair, String)> {
-        use crate::config::KeyringBackend;
-        use keyring::{FileKeyring, Keyring, SystemKeyring, PEER_KEY};
+        use keyring::PEER_KEY;
 
-        let kr: Box<dyn Keyring> = match config.keyring.backend {
-            KeyringBackend::File => {
-                let path = if config.keyring.path.starts_with('/') {
-                    std::path::PathBuf::from(&config.keyring.path)
-                } else {
-                    config.rootdir.join(&config.keyring.path)
-                };
-                let secret =
-                    keyring::load_secret_from_env().map_err(|e| Error::Keyring(e.to_string()))?;
-                Box::new(
-                    FileKeyring::open(&path, secret).map_err(|e| Error::Keyring(e.to_string()))?,
-                )
-            }
-            KeyringBackend::System => {
-                #[cfg(target_os = "linux")]
-                {
-                    if std::env::var("DBUS_SESSION_BUS_ADDRESS").is_ok_and(|v| !v.is_empty()) {
-                        Box::new(SystemKeyring::open(&config.keyring.namespace))
-                    } else if keyring::systemd_creds_available() {
-                        tracing::warn!(
-                            "no D-Bus session available; falling back to systemd-creds keyring"
-                        );
-                        let path = if config.keyring.path.starts_with('/') {
-                            std::path::PathBuf::from(&config.keyring.path)
-                        } else {
-                            config.rootdir.join(&config.keyring.path)
-                        };
-                        let kr = keyring::SystemdCredsKeyring::open(&path)
-                            .map_err(|e| Error::Keyring(e.to_string()))?;
-                        Box::new(kr)
-                    } else {
-                        return Err(Error::Keyring(
-                            "no system keyring available: no D-Bus session for secret-service \
-                             and systemd-creds not found; use --keyring-backend file"
-                                .to_string(),
-                        ));
-                    }
-                }
-                #[cfg(not(target_os = "linux"))]
-                {
-                    Box::new(SystemKeyring::open(&config.keyring.namespace))
-                }
-            }
-            KeyringBackend::SystemdCreds => {
-                #[cfg(target_os = "linux")]
-                {
-                    let path = if config.keyring.path.starts_with('/') {
-                        std::path::PathBuf::from(&config.keyring.path)
-                    } else {
-                        config.rootdir.join(&config.keyring.path)
-                    };
-                    let kr = keyring::SystemdCredsKeyring::open(&path)
-                        .map_err(|e| Error::Keyring(e.to_string()))?;
-                    Box::new(kr)
-                }
-                #[cfg(not(target_os = "linux"))]
-                {
-                    return Err(Error::Keyring(
-                        "systemd-creds backend is only available on Linux".to_string(),
-                    ));
-                }
-            }
-        };
+        let kr = crate::commands::open_keyring(config)?;
 
         // Try to load existing peer key
         match kr.get(PEER_KEY) {
